Relocate lonely mapping utils into `corehq.apps.es.utils`
